requirements.txt
# Dependencies based on the following pip install Dec 12, 2022
# pip install icecream scikit-learn pymoo==0.6.0 pyyaml boto3 pandas scikit-optimize openstudio==3.2.1 tqdm numpy_financial docker openpyxl podman pypiwin32 pip_system_certs click cloudpathlib colorama pyfiglet
about-time==3.1.1
alive-progress==2.4.1
asttokens==2.2.1
autograd==1.5
boto3==1.26.28
botocore==1.29.28
certifi==2022.12.7
charset-normalizer==2.1.1
cma==3.2.2
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
Deprecated==1.2.13
dill==0.3.6
docker==6.0.1
docker-pycreds==0.4.0
et-xmlfile==1.1.0
executing==1.2.0
fonttools==4.38.0
future==0.18.2
grapheme==0.6.0
icecream==2.1.3
idna==3.4
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.5
numpy-financial==1.0.0
openpyxl==3.0.10
openstudio==3.2.1
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pyaml==21.10.1
Pygments==2.13.0
pymoo==0.6.0
pyparsing==3.0.9
pypiwin32==223; sys_platform == 'win32'
python-dateutil==2.8.2
pytz==2022.6
pywin32==305; sys_platform == 'win32'
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
scikit-learn==1.2.0
scikit-optimize==0.9.0
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
tqdm==4.64.1
urllib3==1.26.13
websocket-client==1.4.2
wrapt==1.14.1
podman==4.3.0
pyxdg==0.28
toml==0.10.2
pip_system_certs


